fix issue

+++ src/main/java/mate/academy/springboot/swagger/repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package mate.academy.springboot.swagger.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import mate.academy.springboot.swagger.model.Product;
 import org.springframework.data.domain.PageRequest;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
     @Query("from Product p where  p.price between ?1 and ?2")
-    List<Product> findAllByPriceBetween(Double from, Double to, PageRequest pageRequest);
+    List<Product> findAllByPriceBetween(BigDecimal from, BigDecimal to, PageRequest pageRequest);
 
 }
